add veAllocation.sol's AllocationSetMultiple
