+++ src/mappings/veAllocate.ts
@@ -1,13 +1,9 @@
-import { AllocationSet } from '../@types/veAllocate/veAllocate'
-
-import { veAllocationUpdateType } from './utils/constants'
-
 import {
-  getveAllocateUser,
-  getveAllocateId,
-  getveAllocation,
-  writeveAllocationUpdate
-} from './utils/veUtils'
+  AllocationSet,
+  AllocationSetMultiple
+} from '../@types/veAllocate/veAllocate'
+
+import { handleOneAllocation } from './utils/veUtils'
 
 export function handleAllocationSet(event: AllocationSet): void {
   // get allocation entities
@@ -15,42 +11,25 @@ export function handleAllocationSet(event: AllocationSet): void {
   const nftAddress = event.params.nft.toHexString()
   const chainId = event.params.chainId
   const allocationAmount = event.params.amount.toBigDecimal()
-  const eventId = nftAddress + '-' + chainId.toString()
 
-  const allocateUser = getveAllocateUser(event, eventSender)
-  const allocateId = getveAllocateId(event, eventId)
-  const veAllocation = getveAllocation(event, eventSender, eventId)
+  handleOneAllocation(eventSender, nftAddress, chainId, allocationAmount, event)
+}
 
-  // Update user allocation
-  const newUserAllocation = allocateUser.allocatedTotal.minus(
-    veAllocation.allocated
-  )
-  allocateUser.allocatedTotal = newUserAllocation.plus(allocationAmount)
-
-  // Update id allocation
-  const newIdAllocation = allocateId.allocatedTotal.minus(
-    veAllocation.allocated
-  )
-  allocateId.allocatedTotal = newIdAllocation.plus(allocationAmount)
-
-  veAllocation.allocated = allocationAmount
-  veAllocation.chainId = chainId
-  veAllocation.nftAddress = nftAddress
-
-  allocateUser.lastContact = event.block.timestamp.toI32()
-  allocateId.lastContact = event.block.timestamp.toI32()
-  veAllocation.lastContact = event.block.timestamp.toI32()
-
-  // register allocation update event
-  writeveAllocationUpdate(
-    event,
-    veAllocation.id,
-    veAllocationUpdateType.SET,
-    allocationAmount
-  )
-
-  // save entities
-  allocateUser.save()
-  allocateId.save()
-  veAllocation.save()
+export function handleAllocationSetMultiple(
+  event: AllocationSetMultiple
+): void {
+  // loop
+  for (let i = 0; i < event.params.nft.length; i++) {
+    const eventSender = event.params.sender.toHexString()
+    const nftAddress = event.params.nft[i].toHexString()
+    const chainId = event.params.chainId[i]
+    const allocationAmount = event.params.amount[i].toBigDecimal()
+    handleOneAllocation(
+      eventSender,
+      nftAddress,
+      chainId,
+      allocationAmount,
+      event
+    )
+  }
 }
